pyproject.toml
[tool.poetry]
name = "class-912233-final-sentiment-analysis-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
textblob = "^0.15.3"
tweepy = "^3.10.0"
flask = "^1.1.2"
python-decouple = "^3.4"
replit = "^3.0.0"
flake8 = "^3.9.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
